Added ^0.3 version pin for amphp/process because ^0.2 is buggy.
This addresses downstream issues discovered in Provider/Steam where
amphp/process would fail to launch PHP subprocesses. On Windows this
manifested itself as "'C:\PHP' is not recognized as an internal or
external command, operable program or batch file.", but it also
consistently failed in Travis' Linux environment with a
`Symfony\Component\Process\Exception\ProcessTimedOutException`.